Allow `set_fee_asset` to replace fee_asset_id on all ops.
This is for the `bitshares.transfer` and related methods.
The interface in those just creates all operations with
`fee: { asset_id: 1.3.0 }` and expects `bitshares.finalizeOp`,
to reset the asset_id (if it was provided in kwargs).

`finalizeOp` eventually invokes `transactions.addRequiredFees`,
which *should* replace the `fee_id`.

It can't do this (on my branch, at least) because `addRequiredFees`
keeps old asset_id, assuming it was specifically set by user. (See
the `preset` hack).

This patch is safe and passes all tests. The `set_fee_asset(force=True)`
mode is *only* called from `finalizeOp`, *only* when `fee_asset`
argument is provided. If it *is* provided, the intent *is* to replace
all fee_ids, and that is what's being done in that case. In all other
situations, `set_fee_asset()` is called with `force=False`, this
should be considered the default behavior.